Rename avatar component
