+++ static/themes/common/base.css
@@ -878,7 +878,7 @@
 }
 
 /* ---- Modern Layout Styling ---- */
-.kiwi-messagelist-avatar {
+.kiwi-avatar {
     background-color: var(--brand-default-fg);
     color: var(--brand-default-bg);
     border-color: var(--brand-primary);
