+++ static/themes/dark/theme.css
@@ -122,7 +122,7 @@ button.u-button.u-button-primary.u-submit.kiwi-welcome-znc-start {
     color: #fff;
 }
 
-.kiwi-messagelist-avatar {
+.kiwi-avatar {
     background-color: var(--brand-default-bg);
     color: var(--brand-default-fg);
 }
