+++ static/themes/nightswatch/theme.css
@@ -221,7 +221,7 @@
     border-left-color: var(--brand-midtone);
 }
 
-.kiwi-messagelist-avatar {
+.kiwi-avatar {
     background-color: var(--brand-default-bg);
     color: var(--brand-default-fg);
 }
